Problem with Optional Beacons
Working on the issue that optional beacons can't prematurely exit the
block.  I suppose I could just return at the end, but that wouldn't
teach me anything, would it?

Also, solved a PhantomJS crashing issue by increasing timeouts.  Don't
like the solution, and it means that we have a lot of JS errors in the
usatoday.com page, but it pushes the ball closer to the finish line.